+++ api/endpoints.js
@@ -6,15 +6,10 @@
     tagsUrl:       "http://catalogue/tags",
     cartsUrl:      "http://cart/carts",
     ordersUrl:     "http://orders",
-    customersUrl:  "http://accounts/customers",
-    addressUrl:    "http://accounts/addresses",
-    cardsUrl:      "http://accounts/cards",
-    loginUrl:      "http://login/login",
-    registerUrl:   "http://login/register",
-    userCustUrl:   "http://user/customers",
-    userAddrUrl:   "http://user/addresses",
-    userCardUrl:   "http://user/cards",
-    userLogUrl:    "http://user/login",
-    userRegUrl:    "http://user/register",
+    customersUrl:  "http://user/customers",
+    addressUrl:    "http://user/addresses",
+    cardsUrl:      "http://user/cards",
+    loginUrl:      "http://user/login",
+    registerUrl:   "http://user/register",
   };
 }());
