+++ api/user/index.js
@@ -7,26 +7,47 @@
     , helpers   = require("../../helpers")
     , app       = express()
 
-  app.get("/user/customers/:id", function (req, res, next) {
-    helpers.simpleHttpRequest(endpoints.userCustUrl + "/" + req.params.id, res, next);
+
+
+  app.get("/customers/:id", function (req, res, next) {
+    helpers.simpleHttpRequest(endpoints.customersUrl + "/" + req.params.id, res, next);
   });
-  app.get("/user/cards/:id", function (req, res, next) {
-    helpers.simpleHttpRequest(endpoints.userCardUrl + "/" + req.params.id, res, next);
+  app.get("/cards/:id", function (req, res, next) {
+    helpers.simpleHttpRequest(endpoints.cardsUrl + "/" + req.params.id, res, next);
   });
 
-  app.get("/user/customers", function (req, res, next) {
-    helpers.simpleHttpRequest(endpoints.userCustUrl, res, next);
+  app.get("/customers", function (req, res, next) {
+    helpers.simpleHttpRequest(endpoints.customersUrl, res, next);
   });
-  app.get("/user/addresses", function (req, res, next) {
-    helpers.simpleHttpRequest(endpoints.userAddrUrl, res, next);
+  app.get("/addresses", function (req, res, next) {
+    helpers.simpleHttpRequest(endpoints.addressUrl, res, next);
   });
-  app.get("/user/cards", function (req, res, next) {
-    helpers.simpleHttpRequest(endpoints.userCardUrl, res, next);
+  app.get("/cards", function (req, res, next) {
+    helpers.simpleHttpRequest(endpoints.cardsUrl, res, next);
   });
   // Create Customer - TO BE USED FOR TESTING ONLY (for now)
-  app.post("/user/customers", function(req, res, next) {
+  app.post("/customers", function(req, res, next) {
     var options = {
-      uri: endpoints.userCustUrl,
+      uri: endpoints.customersUrl,
+      method: 'POST',
+      json: true,
+      body: req.body
+    };
+
+    console.log("Posting Customer: " + JSON.stringify(req.body));
+
+    request(options, function (error, response, body) {
+      if (error) {
+        return next(error);
+      }
+      helpers.respondSuccessBody(res, JSON.stringify(body));
+    }.bind({res: res}));
+  });
+
+  // Create Customer - TO BE USED FOR TESTING ONLY (for now)
+  app.post("/register", function(req, res, next) {
+    var options = {
+      uri: endpoints.registerUrl,
       method: 'POST',
       json: true,
       body: req.body
@@ -43,9 +64,9 @@
   });
 
   // Create Address - TO BE USED FOR TESTING ONLY (for now)
-  app.post("/user/addresses", function(req, res, next) {
+  app.post("/addresses", function(req, res, next) {
     var options = {
-      uri: endpoints.userAddrUrl,
+      uri: endpoints.addressUrl,
       method: 'POST',
       json: true,
       body: req.body
@@ -60,9 +81,9 @@
   });
 
   // Create Card - TO BE USED FOR TESTING ONLY (for now)
-  app.post("/user/cards", function(req, res, next) {
+  app.post("/cards", function(req, res, next) {
     var options = {
-      uri: endpoints.userCardUrl,
+      uri: endpoints.cardsUrl,
       method: 'POST',
       json: true,
       body: req.body
@@ -77,10 +98,10 @@
   });
 
   // Delete Customer - TO BE USED FOR TESTING ONLY (for now)
-  app.delete("/user/customers/:id", function(req, res, next) {
+  app.delete("/customers/:id", function(req, res, next) {
     console.log("Deleting Customer " + req.params.id);
     var options = {
-      uri: endpoints.userCustUrl + "/" + req.params.id,
+      uri: endpoints.customersUrl + "/" + req.params.id,
       method: 'DELETE'
     };
     request(options, function (error, response, body) {
@@ -92,10 +113,10 @@
   });
 
   // Delete Address - TO BE USED FOR TESTING ONLY (for now)
-  app.delete("/user/addresses/:id", function(req, res, next) {
+  app.delete("/addresses/:id", function(req, res, next) {
     console.log("Deleting Address " + req.params.id);
     var options = {
-      uri: endpoints.userAddrUrl + "/" + req.params.id,
+      uri: endpoints.addressUrl + "/" + req.params.id,
       method: 'DELETE'
     };
     request(options, function (error, response, body) {
@@ -107,10 +128,10 @@
   });
 
   // Delete Card - TO BE USED FOR TESTING ONLY (for now)
-  app.delete("/user/cards/:id", function(req, res, next) {
+  app.delete("/cards/:id", function(req, res, next) {
     console.log("Deleting Card " + req.params.id);
     var options = {
-      uri: endpoints.userCardUrl + "/" + req.params.id,
+      uri: endpoints.cardsUrl + "/" + req.params.id,
       method: 'DELETE'
     };
     request(options, function (error, response, body) {
@@ -121,7 +142,7 @@
     }.bind({res: res}));
   });
 
-  app.get("/user/login", function (req, res, next) {
+  app.get("/login", function (req, res, next) {
     console.log("Received login request");
 
     async.waterfall([
@@ -130,7 +151,7 @@
             headers: {
               'Authorization': req.get('Authorization')
             },
-            uri: endpoints.userLogUrl
+            uri: endpoints.loginUrl
           };
           request(options, function (error, response, body) {
             if (error) {
