+++ server.js
@@ -8,9 +8,7 @@ var request      = require("request")
   , session      = require("express-session")
   , config       = require("./config")
   , helpers      = require("./helpers")
-  , login        = require("./api/login")
   , cart         = require("./api/cart")
-  , accounts     = require("./api/accounts")
   , catalogue    = require("./api/catalogue")
   , orders       = require("./api/orders")
   , user         = require("./api/user")
@@ -35,9 +33,7 @@ process.argv.forEach(function (val, index, array) {
 });
 
 /* Mount API endpoints */
-app.use(login);
 app.use(cart);
-app.use(accounts);
 app.use(catalogue);
 app.use(orders);
 app.use(user);
